Fix timestamp type
Type of timestamp_ in GroundTruthPackage class changed to long long to
avoid overflow

+++ project_6_extended_kalman_filter/src/ground_truth_package.h
@@ -5,7 +5,7 @@
 
 class GroundTruthPackage {
 public:
-  long timestamp_;
+  long long timestamp_;
 
   enum SensorType{
     LASER,
